Create Weather tool.python
To create a tool that generates weather reports, we'll need to utilize a weather API to fetch weather data and present it to the user. In this example, I'll use the OpenWeatherMap API to fetch weather data for a given location. You'll need to sign up for an API key from OpenWeatherMap (https://openweathermap.org/api) to use their service.

Signed-off-by: shivangi241103 <115358526+shivangi241103@users.noreply.github.com>